fix(auth): harden email verification fallback
