+++ api/src/api/http/auth.rs
@@ -2299,7 +2299,7 @@ pub async fn verify_email_get(
                     // Not an outcome yet: the interactive POST records the terminal one.
                     emit("accepted", Some("interactive_handoff"), 303).await;
                     let interactive_url = format!(
-                        "/verify-email/interactive?token={}",
+                        "/email-verification/continue?token={}",
                         urlencoding::encode(&token)
                     );
                     Redirect::to(&interactive_url).into_response()
@@ -2506,19 +2506,20 @@ fn verify_html_page_with_refresh(
     let body = format!(
         "<!doctype html><html lang=\"en\"><head><meta charset=\"utf-8\">{refresh_tag}\
          <meta name=\"viewport\" content=\"width=device-width, initial-scale=1\">\
-         <meta name=\"theme-color\" content=\"#072218\"><title>{heading} - Divine</title>\
-         <style>*{{box-sizing:border-box}}body{{font-family:Inter,system-ui,-apple-system,sans-serif;\
+         <meta name=\"theme-color\" content=\"#072218\"><title>{heading} - {brand}</title>\
+         <style>*{{box-sizing:border-box}}body{{font-family:system-ui,-apple-system,sans-serif;\
          background:#072218;color:#f7fffc;display:flex;min-height:100vh;align-items:center;\
          justify-content:center;margin:0;padding:1rem}}.card{{width:100%;max-width:420px;text-align:center;\
          background:#0f2e23;border:1px solid rgba(39,197,139,.25);border-radius:1rem;padding:2rem;\
          box-shadow:0 18px 60px rgba(0,0,0,.22)}}.brand{{display:inline-flex;flex-direction:column;\
          align-items:center;gap:2px;margin-bottom:1.5rem;text-decoration:none}}.brand img{{height:28px;\
          max-width:180px}}.brand span{{color:#27c58b;font-size:11px;font-weight:600;letter-spacing:3px;\
-         text-transform:uppercase}}h1{{font-family:'Bricolage Grotesque',Inter,system-ui,sans-serif;\
+         text-transform:uppercase}}h1{{font-family:system-ui,-apple-system,sans-serif;\
          font-size:1.55rem;line-height:1.2;margin:0 0 .75rem}}p{{color:#b8cbc4;margin:0;\
          font-size:.98rem;line-height:1.55}}</style></head><body><main class=\"card\">\
-         <a class=\"brand\" href=\"/\"><img src=\"/divine-logo.svg\" alt=\"Divine\">\
+         <a class=\"brand\" href=\"/\"><img src=\"/divine-logo.svg\" alt=\"{brand}\">\
          <span>Login</span></a><h1>{heading}</h1><p>{message}</p></main></body></html>",
+        brand = html_escape(BRAND_NAME),
         heading = html_escape(heading),
         message = html_escape(message),
     );
@@ -5368,6 +5369,7 @@ mod tests {
     use super::validate_origin;
     use super::verify_html_page;
     use super::AccountStatusResponse;
+    use super::BRAND_NAME;
     #[cfg(feature = "integration-tests")]
     use super::{VERIFICATION_LINK_SUPERSEDED_CODE, VERIFICATION_LINK_SUPERSEDED_HEADING};
     use axum::response::IntoResponse;
@@ -5389,6 +5391,9 @@ mod tests {
         assert!(body.contains(">Login</span>"));
         assert!(body.contains("background:#072218"));
         assert!(body.contains("Email verified!"));
+        assert!(body.contains(&format!("Email verified! - {BRAND_NAME}")));
+        assert!(!body.contains("Bricolage Grotesque"));
+        assert!(!body.contains("font-family:Inter"));
     }
 
     #[tokio::test]
@@ -5628,7 +5633,7 @@ mod tests {
         VerifyEmailQuery, VerifyEmailRequest,
     };
     #[cfg(feature = "integration-tests")]
-    use crate::api::http::routes::AuthState;
+    use crate::api::http::routes::{public_verify_email_route, AuthState};
     #[cfg(feature = "integration-tests")]
     use crate::api::tenant::{Tenant, TenantExtractor};
     #[cfg(feature = "integration-tests")]
@@ -5639,10 +5644,11 @@ mod tests {
     use crate::state::KeycastState;
     #[cfg(feature = "integration-tests")]
     use axum::{
+        body::Body,
         extract::{Query, State},
         http::{
             header::{AUTHORIZATION, ORIGIN},
-            HeaderMap, HeaderValue, StatusCode,
+            HeaderMap, HeaderValue, Request, StatusCode,
         },
         Json,
     };
@@ -5663,6 +5669,8 @@ mod tests {
     use sqlx::PgPool;
     #[cfg(feature = "integration-tests")]
     use std::sync::Arc;
+    #[cfg(feature = "integration-tests")]
+    use tower::ServiceExt;
     #[cfg(feature = "integration-tests")]
     use uuid::Uuid;
     #[cfg(feature = "integration-tests")]
@@ -6548,11 +6556,39 @@ mod tests {
             .and_then(|value| value.to_str().ok())
             .unwrap();
         assert!(
-            location.starts_with("/verify-email/interactive?token="),
+            location.starts_with("/email-verification/continue?token="),
             "first-party tokens must be handed to the interactive page, got {location}"
         );
 
         cleanup_verify_email_test_data(&pool, &pubkey, &token).await;
+    }
+
+    #[cfg(feature = "integration-tests")]
+    #[tokio::test]
+    async fn test_public_verify_email_route_is_registered_without_cors() {
+        let pool = create_test_db().await;
+        let auth_state = create_test_auth_state(pool);
+        let app = public_verify_email_route(auth_state.state, auth_state.auth_tx);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/verify-email?token=sensitive")
+                    .header(ORIGIN, "https://example.com")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .expect("public verification route response");
+
+        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+        assert!(!response
+            .headers()
+            .contains_key(axum::http::header::ACCESS_CONTROL_ALLOW_ORIGIN));
+        let body = axum::body::to_bytes(response.into_body(), usize::MAX)
+            .await
+            .expect("route response body");
+        assert_eq!(&body[..], b"Missing Host header");
     }
 
     /// A token that resolves to nothing - already used, or superseded by a newer verification
