--- web/src/routes/verify-email/interactive/+page.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import VerifyEmail from '../+page.svelte';
-</script>
-
-<VerifyEmail />
